+++ src/excmds.ts
@@ -31,7 +31,7 @@
 
     A "splat" operator (...) means that the excmd will accept any number of space-delimited arguments into that parameter.
 
-    You do not need to worry about types.
+    You do not need to worry about types. Return values which are promises will turn into whatever they promise to when used in [[composite]].
 
     At the bottom of each function's help page, you can click on a link that will take you straight to that function's definition in our code. This is especially recommended for browsing the [config](/static/docs/modules/_config_.html#defaults) which is nigh-on unreadable on these pages.
 
@@ -57,14 +57,15 @@
 
     There are some caveats common to all webextension vimperator-alikes:
 
-    - Do not try to navigate to any about:\* pages using `:open` as it will
-      fail silently
-    - Firefox will not load Tridactyl on addons.mozilla.org, about:\*, some
-      file:\* URIs, view-source:\*, or data:\*. On these pages Ctrl-L (or F6),
-      Ctrl-Tab and Ctrl-W are your escape hatches
-    - Tridactyl does not currently support changing/hiding the Firefox GUI, but
-      you can do it yourself by changing your userChrome. There is an [example
-      file](2) available in our repository.
+    - To make Tridactyl work on addons.mozilla.org, about:\*, some file:\*
+      URIs, view-source:\*, or data:\*, you need to open `about:config`, add a
+      new boolean `privacy.resistFingerprinting.block_mozAddonManager` with the
+      value `true`, and remove the above domains from
+      `extensions.webextensions.restrictedDomains`.
+    - To change/hide the GUI of Firefox from Tridactyl, you can use [[guiset]]
+      with the native messenger installed (see [[native]] and
+      [[installnative]]). Alternatively, you can edit your userChrome yourself.
+      There is an [example file](2) available in our repository.
 
     If you want a more fully-featured vimperator-alike, your best option is
     [Firefox ESR][3] and Vimperator :)
@@ -143,16 +144,24 @@ export async function getNativeVersion(): Promise<void> {
  */
 //#content
 export async function fillinput(...content: string[]) {
-    let inputToFill = DOM.getLastUsedInput() as HTMLInputElement
-    inputToFill.value = content.join(" ")
+    let inputToFill = DOM.getLastUsedInput()
+    if ("value" in inputToFill) {
+        ;(inputToFill as HTMLInputElement).value = content.join(" ")
+    } else {
+        inputToFill.textContent = content.join(" ")
+    }
 }
 
 /** @hidden */
 //#content
 export async function getinput() {
     // this should probably be subsumed by the focusinput code
-    let input = DOM.getLastUsedInput() as HTMLInputElement
-    return input.value
+    let input = DOM.getLastUsedInput()
+    if ("value" in input) {
+        return (input as HTMLInputElement).value
+    } else {
+        return input.textContent
+    }
 }
 
 /**
@@ -167,8 +176,9 @@ export async function getinput() {
  */
 //#background
 export async function editor() {
+    let url = new URL((await activeTab()).url)
     if (!await Native.nativegate()) return
-    const file = (await Native.temp(await getinput())).content
+    const file = (await Native.temp(await getinput(), url.hostname)).content
     fillinput((await Native.editor(file)).content)
     // TODO: add annoying "This message was written with [Tridactyl](https://addons.mozilla.org/en-US/firefox/addon/tridactyl-vim/)"
     // to everything written using editor
@@ -180,7 +190,7 @@ import * as css_util from "./css_util"
 /**
  * Change which parts of the Firefox user interface are shown. **NB: This feature is experimental and might break stuff.**
  *
- * Might mangle your userChrome. Requires native messenger, and you must restart Firefox each time to see any changes. <!-- (unless you enable addon debugging and refresh using the browser toolbox) -->
+ * Might mangle your userChrome. Requires native messenger, and you must restart Firefox each time to see any changes (this can be done using [[restart]]). <!-- (unless you enable addon debugging and refresh using the browser toolbox) -->
  *
  * View available rules and options [here](/static/docs/modules/_css_util_.html#potentialrules) and [here](/static/docs/modules/_css_util_.html#metarules).
  *
@@ -253,6 +263,23 @@ export function cssparse(...css: string[]) {
 }
 
 /**
+ *
+ * Simply sets
+ * ```js
+ *  "privacy.resistFingerprinting.block_mozAddonManager":true
+ *  "extensions.webextensions.restrictedDomains":""
+ * ```
+ * in about:config via user.js so that Tridactyl (and other extensions!) can be used on addons.mozilla.org and other sites.
+ *
+ * Requires `native`.
+ */
+//#background
+export async function fixamo() {
+    await Native.writePref("privacy.resistFingerprinting.block_mozAddonManager", true)
+    await Native.writePref("extensions.webextensions.restrictedDomains", "")
+}
+
+/**
  * Uses the native messenger to open URLs.
  *
  * **Be *seriously* careful with this: you can use it to open any URL you can open in the Firefox address bar.**
@@ -288,7 +315,7 @@ export async function exclaim(...str: string[]) {
  */
 //#background
 export async function exclaim_quiet(...str: string[]) {
-    ;(await Native.run(str.join(" "))).content
+    return (await Native.run(str.join(" "))).content
 }
 
 /**
@@ -366,6 +393,21 @@ export async function updatenative(interactive = true) {
 
         if (interactive) native()
     }
+}
+
+/**
+ *  Restarts firefox with the same commandline arguments.
+ *
+ *  Warning: This can kill your tabs, especially if you :restart several times
+ *  in a row
+ */
+//#background
+export async function restart() {
+    const firefox = (await Native.ffargs()).join(" ")
+    const profile = await Native.getProfileDir()
+    // Wait for the lock to disappear, then wait a bit more, then start firefox
+    Native.run(`while readlink ${profile}/lock ; do sleep 1 ; done ; sleep 1 ; ${firefox}`)
+    qall()
 }
 
 // }}}
@@ -1586,15 +1628,36 @@ export function repeat(n = 1, ...exstr: string[]) {
     for (let i = 0; i < n; i++) controller.acceptExCmd(cmd)
 }
 
-/** Split `cmds` on pipes (|) and treat each as its own command.
-
-    Workaround: this should clearly be in the parser, but we haven't come up with a good way to deal with |s in URLs, search terms, etc. yet.
-*/
+/**
+ * Split `cmds` on pipes (|) and treat each as its own command. Return values are cast to strings and passed to the appended to the arguments of the next ex command, e.g,
+ *
+ * `composite echo yes | fillcmdline` becomes `fillcmdline yes`. A more complicated example is the ex alias, `command current_url composite get_current_url | fillcmdline_notrail `, which is used in, e.g. `bind T current_url tabopen`.
+ *
+ * Workaround: this should clearly be in the parser, but we haven't come up with a good way to deal with |s in URLs, search terms, etc. yet.
+ *
+ * `cmds` are also split with semicolons (;) and don't pass things along to each other.
+ *
+ * The behaviour of combining ; and | in the same composite command is left as an exercise for the reader.
+ */
 //#background
 export async function composite(...cmds: string[]) {
     cmds = cmds.join(" ").split("|")
+    let val = ""
     for (let c of cmds) {
-        await controller.acceptExCmd(c)
+        let dmds = c.split(";")
+        if (dmds.length > 1) {
+            for (let d of dmds) {
+                await controller.acceptExCmd(d)
+            }
+        } else val = await controller.acceptExCmd(dmds[0] + val)
+        try {
+            if (val == undefined || val.includes("undefined")) val = ""
+            else val = " " + val
+        } catch (e) {
+            if (e instanceof TypeError) {
+                val = " " + val
+            } else throw e
+        }
     }
 }
 
@@ -1626,13 +1689,12 @@ export function fillcmdline_notrail(...strarr: string[]) {
     messageActiveTab("commandline_frame", "fillcmdline", [str, trailspace])
 }
 
-/** Equivalent to `fillcmdline_notrail <yourargs><current URL>`
-
-    See also [[fillcmdline_notrail]]
-*/
+/**
+ * Returns the current URL. For use with [[composite]].
+ */
 //#background
-export async function current_url(...strarr: string[]) {
-    fillcmdline_notrail(...strarr, (await activeTab()).url)
+export async function get_current_url() {
+    return (await activeTab()).url
 }
 
 /** Use the system clipboard.
@@ -1871,9 +1933,15 @@ export function set(key: string, ...values: string[]) {
 
 /** Set autocmds to run when certain events happen.
 
- @param event Curently, only 'DocStart' is supported.
+ @param event Curently, only 'TriStart' and 'DocStart' are supported.
 
- @param url The URL on which the events will trigger (currently just uses "contains")
+ @param url For DocStart: the URL on which the events will trigger (currently
+ just uses "contains").
+
+ For TriStart: A regular expression that matches the hostname of the computer
+ the autocmd should be run on. This requires the native messenger to be
+ installed, except for the ".*" regular expression which will always be
+ triggered, even without the native messenger.
 
  @param excmd The excmd to run (use [[composite]] to run multiple commands)
 
@@ -1882,7 +1950,7 @@ export function set(key: string, ...values: string[]) {
 export function autocmd(event: string, url: string, ...excmd: string[]) {
     // rudimentary run time type checking
     // TODO: Decide on autocmd event names
-    if (!["DocStart"].includes(event)) throw event + " is not a supported event."
+    if (!["DocStart", "TriStart"].includes(event)) throw event + " is not a supported event."
     config.set("autocmds", event, url, excmd.join(" "))
 }
 
@@ -2318,6 +2386,32 @@ export async function bmark(url?: string, ...titlearr: string[]) {
     }
 
     browser.bookmarks.create({ url, title })
+}
+
+//#background
+export async function echo(...str: string[]) {
+    return str.join(" ")
+}
+
+/**
+ * Lets you execute JavaScript in the page context. If you want to get the result back, use `composite js ... | fillcmdline`
+ *
+ * Some of Tridactyl's functions are accessible here via the `tri` object. Just do `console.log(tri)` in the web console on the new tab page to see what's available.
+ *
+ * Aliased to `!js`
+ *
+ */
+//#content
+export async function js(...str: string[]) {
+    return eval(str.join(" "))
+}
+
+/**
+ * Lets you execute JavaScript in the background context. All the help from [[js]] applies. Gives you a different `tri` object.
+ */
+//#background
+export async function jsb(...str: string[]) {
+    return eval(str.join(" "))
 }
 
 /**  Open a welcome page on first install.