+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Tridactyl changelog
 
+## Release 1.13.0 / Unreleased
+
+- `fixamo` added: you can now use Tridactyl on addons.mozilla.org. Requires a `restart`.
+
+- `editor` now includes the hostname of the site you are on in the temporary filename
+    - this is mostly so that you can set up syntax highlighting in Vim, e.g,
+    - `au BufReadPost *github.com* set syntax=pandoc`
+
+- `native` support for Windows: just do what `installnative` tells you to
+    - you'll probably want to make sure `gvim` is on your path
+
+- **Potentially breaking change**: pipes in `composite` now send return values to the following ex command. Use semi-colons if you want the old behaviour back (see `bind D`).
+
+- Add internal functions for editing `user.js` - we'll probably add a nice interface to this some day.
+
+- Add `restart` command to restart Firefox.
+
 ## Release 1.12.0 / 2018-05-13
 
 - Add container support